refactor: reduce Next.js extractor bumpy road
